requirements.txt
streamlit
numpy
Pillow
opencv-python-headless
matplotlib
scikit-learn
joblib
tensorflow
xgboost
